pull to main branch from Kafka connect

+++ ExamAdministrationService/src/main/java/com/examManagement/ExamAdministrationService/service/ExamServiceImpl.java
@@ -8,11 +8,16 @@ import com.examManagement.ExamAdministrationService.entity.ExamStatus;
 import com.examManagement.ExamAdministrationService.exception.ResourceNotFoundException;
 import com.examManagement.ExamAdministrationService.repository.ExamRepository;
 import com.examManagement.ExamAdministrationService.repository.ExaminerRepository;
+import com.examManagement.ExamAdministrationService.validator.ExamValidator;
 import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.CachePut;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -24,6 +29,8 @@ public class ExamServiceImpl implements ExamService{
     private final ExamRepository examRepository;
     private final ExaminerService examinerService;
     private final ExaminerRepository examinerRepository;
+    private final KafkaTemplate<String, String> kafkaTemplate;
+    private final ExamValidator examValidator;
 
     @Override
     public ExamResponse createExam(ExamRequest request) {
@@ -40,13 +47,18 @@ public class ExamServiceImpl implements ExamService{
                 .collect(Collectors.toList());
     }
 
+    @Cacheable(value = "exam-info", key = "#id",
+            unless = "#result == null")
     @Override
     public ExamResponse getExamById(String id) {
         Exam exam = examRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Exam not found"));
+        System.out.println("👉 GỌI EXAM SERVICE (CACHE): " + id);
         return ExamMapper.toResponse(exam);
     }
 
+    @CachePut(value = "exam-info", key = "#id",
+            unless = "#result == null")
     @Override
     public ExamResponse updateExam(String id, ExamRequest request) {
         Exam updatedExam = examRepository.findById(id)
@@ -56,15 +68,33 @@ public class ExamServiceImpl implements ExamService{
         updatedExam.setScheduledTime(request.getScheduledTime());
         updatedExam.setName(request.getName());
         updatedExam.setDuration(request.getDuration());
-        updatedExam.setAssignedExaminerId(request.getAssignedExaminerId());
+        updateAssignedExaminer(updatedExam, request.getAssignedExaminerId());
         examRepository.save(updatedExam);
         return ExamMapper.toResponse(updatedExam);
+    }
+
+    @CachePut(value = "exam-info", key = "#exam.id",
+            unless = "#result == null")
+    private void updateAssignedExaminer(Exam exam, List<String> examinerIds){
+        for (String examinerId: examinerIds) {
+            if (!examinerRepository.existsById(examinerId)) {
+                throw new ResourceNotFoundException("Examiner not found with ID: " + examinerId);
+            }
+            if (!examinerRepository.existsByIdAndActiveTrue(examinerId)){
+                throw new IllegalStateException("Cannot assign inactive examiner");
+            }
+        }
+        if (!new HashSet<>(examinerIds).equals(new HashSet<>(exam.getAssignedExaminerId()))) {
+            exam.setAssignedExaminerId(new ArrayList<>(examinerIds));
+        }
     }
 
     @Override
     public void deleteExam(String id) {
         Exam deletedExam = examRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Exam not found"));
+        if (deletedExam.getStatus() == ExamStatus.PUBLISHED)
+            throw new IllegalStateException("Cannot delete published exams can be deleted");
         examRepository.delete(deletedExam);
     }
 
@@ -75,50 +105,12 @@ public class ExamServiceImpl implements ExamService{
                 .collect(Collectors.toList());
     }
 
+    @CachePut(value = "exam-info", key = "#id",
+            unless = "#result == null")
     @Override
-    public ExamResponse addExaminerToExam(String examId, String examinerId) {
-        Exam exam = examRepository.findById(examId)
-                .orElseThrow(() -> new ResourceNotFoundException("Exam not found with ID: " + examId));
-
-        if (!examinerRepository.existsById(examinerId)) {
-            throw new ResourceNotFoundException("Examiner not found with ID: " + examinerId);
-        }
-
-        // Check if examiner is already assigned
-        if (exam.getAssignedExaminerId() != null &&
-                exam.getAssignedExaminerId().contains(examinerId)) {
-            throw new IllegalArgumentException("Examiner is already assigned to this exam");
-        }
-
-        if (exam.getAssignedExaminerId() == null) {
-            exam.setAssignedExaminerId(new ArrayList<>());
-        }
-        exam.getAssignedExaminerId().add(examinerId);
-
-        Exam updatedExam = examRepository.save(exam);
-        return ExamMapper.toResponse(updatedExam);
-    }
-
-    @Override
-    public ExamResponse removeExaminerFromExam(String examId, String examinerId) {
-        Exam exam = examRepository.findById(examId)
-                .orElseThrow(() -> new ResourceNotFoundException("Exam not found with ID: " + examId));
-
-        // Check if examiner is assigned
-        if (exam.getAssignedExaminerId() == null ||
-                !exam.getAssignedExaminerId().contains(examinerId)) {
-            throw new IllegalArgumentException("Examiner is not assigned to this exam");
-        }
-
-        exam.getAssignedExaminerId().remove(examinerId);
-        Exam updatedExam = examRepository.save(exam);
-        return ExamMapper.toResponse(updatedExam);
-    }
-
-    @Override
-    public ExamResponse publishExam(String examId) {
-        Exam exam = examRepository.findById(examId)
-                .orElseThrow(() -> new ResourceNotFoundException("Exam not found: "+examId));
+    public ExamResponse publishExam(String id) {
+        Exam exam = examRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Exam not found: "+id));
 
         if (exam.getStatus() != ExamStatus.DRAFT) {
             throw new IllegalStateException("Only DRAFT exams can be published");
@@ -129,39 +121,54 @@ public class ExamServiceImpl implements ExamService{
         return ExamMapper.toResponse(exam);
     }
 
+    @CachePut(value = "exam-info", key = "#id",
+            unless = "#result == null")
     @Override
-    public ExamResponse completeExam(String examId) {
-        Exam exam = examRepository.findById(examId)
+    public ExamResponse scoreExam(String id) {
+        Exam exam = examRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Exam not found"));
+        examValidator.validateExamStatusForScoring(exam);
+
+        exam.setStatus(ExamStatus.SCORED);
+        examRepository.save(exam);
+        return ExamMapper.toResponse(exam);
+    }
+
+    @CachePut(value = "exam-info", key = "#id",
+            unless = "#result == null")
+    @Override
+    public ExamResponse completeExam(String id) {
+        Exam exam = examRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Exam not found"));
 
-        if (exam.getStatus() != ExamStatus.PUBLISHED) {
-            throw new IllegalStateException("Only PUBLISHED exams can be completed");
-        }
+        examValidator.validateExamStatusForCompletion(exam);
 
+        sendKafkaEvent(id, "exam-complete-events");
         exam.setStatus(ExamStatus.COMPLETED);
         examRepository.save(exam);
         return ExamMapper.toResponse(exam);
     }
 
+    @CachePut(value = "exam-info", key = "#id",
+            unless = "#result == null")
     @Override
-    public ExamResponse cancelExam(String examId) {
-        Exam exam = examRepository.findById(examId)
+    public ExamResponse cancelExam(String id) {
+        Exam exam = examRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Exam not found"));
 
-        if (exam.getStatus() != ExamStatus.PUBLISHED) {
-            throw new IllegalStateException("Only PUBLISHED exams can be canceled");
-        }
+        examValidator.validateExamStatusForCancellation(exam);
 
         exam.setStatus(ExamStatus.CANCELED);
         examRepository.save(exam);
         return ExamMapper.toResponse(exam);
     }
 
+    @CachePut(value = "exam-info", key = "#id",
+            unless = "#result == null")
     @Override
-    public ExamResponse revertToDraft(String examId) {
-        Exam exam = examRepository.findById(examId)
+    public ExamResponse revertToDraft(String id) {
+        Exam exam = examRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Exam not found"));
-
         exam.setStatus(ExamStatus.DRAFT);
         examRepository.save(exam);
         return ExamMapper.toResponse(exam);
@@ -171,21 +178,7 @@ public class ExamServiceImpl implements ExamService{
         public List<ExamResponse> batchPublishExams(List<String> examIds) {
             List<Exam> exams = examRepository.findAllById(examIds);
 
-            if (exams.size() != examIds.size()) {
-                List<String> foundIds = exams.stream().map(Exam::getId).toList();
-                List<String> missingIds = examIds.stream()
-                        .filter(id -> !foundIds.contains(id))
-                        .toList();
-                throw new ResourceNotFoundException("Missing exams: " + missingIds);
-            }
-
-            exams.forEach(exam -> {
-                if (exam.getStatus() != ExamStatus.DRAFT) {
-                    throw new IllegalStateException(
-                            "Exam " + exam.getId() + " is not in DRAFT state"
-                    );
-                }
-            });
+            examValidator.validateExamsForBatchPublish(exams, examIds);
 
             exams.forEach(exam -> exam.setStatus(ExamStatus.PUBLISHED));
             examRepository.saveAll(exams);
@@ -195,4 +188,12 @@ public class ExamServiceImpl implements ExamService{
                     .toList();
         }
 
+    private void sendKafkaEvent(String examId, String topic) {
+        try {
+            kafkaTemplate.send(topic, examId);
+            System.out.println("✅ Đã gửi message vào Kafka topic.");
+        } catch (Exception ex) {
+            System.err.println("❌ Lỗi khi gửi Kafka: " + ex.getMessage());
+        }
+    }
 }
